+++ src/main/java/com/irrigation/erp/backend/dto/InventoryRequestLineItemResponseDTO.java
@@ -5,6 +5,8 @@ import com.irrigation.erp.backend.enums.RequestLineItemStatus;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class InventoryRequestLineItemResponseDTO {
@@ -12,8 +14,8 @@ public class InventoryRequestLineItemResponseDTO {
     private Long requestedItemId;
     private String requestedItemCode;
     private String requestedItemName;
-    private Double requestedQuantity;
-    private Double currentStockQuantity;
+    private BigDecimal requestedQuantity;
+    private BigDecimal currentStockQuantity;
 
     private RequestLineItemStatus status;
